code/requirementsMac.txt
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorlog==4.6.2
elasticsearch==7.10.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
neo4j==4.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
urllib3==1.26.2
Werkzeug==1.0.1
